Make add part in admin page

+++ src/main/java/cat/mood/koryto/model/Car.java
@@ -10,7 +10,7 @@ import lombok.extern.slf4j.Slf4j;
 @AllArgsConstructor
 @Slf4j
 public class Car {
-    Long carId;
+    Integer carId;
     String carBrand;
     String carModel;
 }
